+++ packages/dts-plugin/src/core/lib/utils.ts
@@ -151,8 +151,8 @@ const getEnvHeaders = (): Record<string, string> => {
 };
 
 export async function axiosGet(url: string, config?: AxiosRequestConfig) {
-  const httpAgent = new http.Agent({ family: 4 });
-  const httpsAgent = new https.Agent({ family: 4 });
+  const httpAgent = new http.Agent({ family: config?.family ?? 4 });
+  const httpsAgent = new https.Agent({ family: config?.family ?? 4 });
   return axios.get(url, {
     httpAgent,
     httpsAgent,