+++ packages/sdk/src/types/plugins/ModuleFederationPlugin.ts
@@ -156,6 +156,8 @@ export interface DtsHostOptions {
   runtimePkgs?: string[];
   remoteTypeUrls?: (() => Promise<RemoteTypeUrls>) | RemoteTypeUrls;
   timeout?: number;
+  /** The family of IP, used for network requests */
+  family?: 4 | 6;
   typesOnBuild?: boolean;
 }