+++ packages/create-module-federation/CHANGELOG.md
@@ -1,5 +1,7 @@
 # create-module-federation
 
+## 0.18.0
+
 ## 0.17.1
 
 ### Patch Changes