+++ packages/esbuild/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @module-federation/esbuild
 
+## 0.0.75
+
+### Patch Changes
+
+- Updated dependencies [f6381e6]
+  - @module-federation/sdk@0.18.0
+
 ## 0.0.74
 
 ### Patch Changes