+++ packages/nextjs-mf/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @module-federation/nextjs-mf
 
+## 8.8.35
+
+### Patch Changes
+
+- Updated dependencies [609d477]
+- Updated dependencies [0ab51b8]
+- Updated dependencies [98a29c3]
+- Updated dependencies [f6381e6]
+  - @module-federation/runtime@0.18.0
+  - @module-federation/enhanced@0.18.0
+  - @module-federation/sdk@0.18.0
+  - @module-federation/node@2.7.11
+  - @module-federation/webpack-bundler-runtime@0.18.0
+
 ## 8.8.34
 
 ### Patch Changes