+++ packages/rspack/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @module-federation/rspack
 
+## 0.18.0
+
+### Patch Changes
+
+- Updated dependencies [08f089a]
+- Updated dependencies [f6381e6]
+  - @module-federation/dts-plugin@0.18.0
+  - @module-federation/sdk@0.18.0
+  - @module-federation/runtime-tools@0.18.0
+  - @module-federation/manifest@0.18.0
+  - @module-federation/bridge-react-webpack-plugin@0.18.0
+  - @module-federation/managers@0.18.0
+  - @module-federation/inject-external-runtime-core-plugin@0.18.0
+
 ## 0.17.1
 
 ### Patch Changes