+++ packages/rspress-plugin/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @module-federation/rspress-plugin
 
+## 0.18.0
+
+### Patch Changes
+
+- Updated dependencies [0ab51b8]
+- Updated dependencies [98a29c3]
+- Updated dependencies [f6381e6]
+- Updated dependencies [38b8d24]
+  - @module-federation/enhanced@0.18.0
+  - @module-federation/sdk@0.18.0
+  - @module-federation/rsbuild-plugin@0.18.0
+  - @module-federation/error-codes@0.18.0
+
 ## 0.17.1
 
 ### Patch Changes